increased data api usability
This has a large number of changes in one commit. Sorry.

Session management has been added for long-running scripts or scripts where you are interacting with the data api across a long-term session. The library is now capable of fully-automatically refreshing its own tokens upon expiry. There are also new levers to get and set tokens (for situations where you pass a token between sessions or pages ... and not being able to renew that token is OK) and so on.

Additional docs (README) forthcoming.