Unilaterally relicense the library to BSD-3
Changing to BSD-3. Source code will continue to be made available, and all changes to the library will be upstreamed (to this repository). We will be selling a product which relies on this library, which cannot happen under GPL-3 without the client also publishing their entire frontend source code. For the purpose, that's extremely inappropriate. This license will only apply to the 2.X and newer released version of the library. Versions >1.x<2.y will continue to be distributed as GPL-3, with all the rights and freedoms conferred.